started auto play that's still under construction
