+++ app/src/main/java/com/pursuit/tictactoe/RulesActivity.kt
@@ -5,6 +5,7 @@ import android.content.pm.ActivityInfo
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.animation.AnimationUtils
+import com.pursuit.tictactoe.main.MainActivity
 import kotlinx.android.synthetic.main.activity_rules.*
 
 class RulesActivity : AppCompatActivity() {
@@ -18,6 +19,7 @@ class RulesActivity : AppCompatActivity() {
         play.setOnClickListener {
             startActivity(Intent(this@RulesActivity, MainActivity::class.java))
         }
+
 
     }
