+++ app/src/main/java/com/pursuit/tictactoe/results/Result2Activity.kt
@@ -1,29 +1,31 @@
-package com.pursuit.tictactoe
+package com.pursuit.tictactoe.results
 
 import android.content.Intent
 import android.content.pm.ActivityInfo
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.animation.AnimationUtils
-import kotlinx.android.synthetic.main.activity_result.*
+import com.pursuit.tictactoe.main.MainActivity
+import com.pursuit.tictactoe.R
+import kotlinx.android.synthetic.main.activity_result2.*
 
-class ResultActivity : AppCompatActivity() {
+
+class Result2Activity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         this.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
-        setContentView(R.layout.activity_result)
+        setContentView(R.layout.activity_result2)
 
         startAnimation()
-        yes1Btn.setOnClickListener { startActivity(Intent(this@ResultActivity, MainActivity::class.java)) }
-        no1Btn.setOnClickListener { finishAffinity() }
+        yes2Btn.setOnClickListener { startActivity(Intent(this@Result2Activity, MainActivity::class.java)) }
+        no2Btn.setOnClickListener { finishAffinity() }
 
     }
 
     private fun startAnimation() {
         val animation = AnimationUtils.loadAnimation(applicationContext, R.anim.scale_anim)
-        win.startAnimation(animation)
+        win2.startAnimation(animation)
     }
-
 
 }
