+++ app/src/main/java/com/pursuit/tictactoe/results/ResultActivity.kt
@@ -1,26 +1,31 @@
-package com.pursuit.tictactoe
+package com.pursuit.tictactoe.results
 
 import android.content.Intent
 import android.content.pm.ActivityInfo
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.animation.AnimationUtils
-import kotlinx.android.synthetic.main.activity_tie.*
+import com.pursuit.tictactoe.main.MainActivity
+import com.pursuit.tictactoe.R
+import kotlinx.android.synthetic.main.activity_result.*
 
-class TieActivity : AppCompatActivity() {
+class ResultActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         this.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
-        setContentView(R.layout.activity_tie)
+        setContentView(R.layout.activity_result)
 
         startAnimation()
-        yes3Btn.setOnClickListener { startActivity(Intent(this@TieActivity, MainActivity::class.java)) }
-        no3Btn.setOnClickListener { finishAffinity() }
+        yes1Btn.setOnClickListener { startActivity(Intent(this@ResultActivity, MainActivity::class.java)) }
+        no1Btn.setOnClickListener { finishAffinity() }
+
     }
 
     private fun startAnimation() {
         val animation = AnimationUtils.loadAnimation(applicationContext, R.anim.scale_anim)
-        tie.startAnimation(animation)
+        win.startAnimation(animation)
     }
+
+
 }
