+++ app/src/main/java/com/pursuit/tictactoe/results/TieActivity.kt
@@ -1,29 +1,28 @@
-package com.pursuit.tictactoe
+package com.pursuit.tictactoe.results
 
 import android.content.Intent
 import android.content.pm.ActivityInfo
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.animation.AnimationUtils
-import kotlinx.android.synthetic.main.activity_result2.*
+import com.pursuit.tictactoe.main.MainActivity
+import com.pursuit.tictactoe.R
+import kotlinx.android.synthetic.main.activity_tie.*
 
-
-class Result2Activity : AppCompatActivity() {
+class TieActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         this.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
-        setContentView(R.layout.activity_result2)
+        setContentView(R.layout.activity_tie)
 
         startAnimation()
-        yes2Btn.setOnClickListener { startActivity(Intent(this@Result2Activity, MainActivity::class.java)) }
-        no2Btn.setOnClickListener { finishAffinity() }
-
+        yes3Btn.setOnClickListener { startActivity(Intent(this@TieActivity, MainActivity::class.java)) }
+        no3Btn.setOnClickListener { finishAffinity() }
     }
 
     private fun startAnimation() {
         val animation = AnimationUtils.loadAnimation(applicationContext, R.anim.scale_anim)
-        win2.startAnimation(animation)
+        tie.startAnimation(animation)
     }
-
 }
