+++ app/src/main/res/layout/activity_main.xml
@@ -17,7 +17,7 @@
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintBottom_toTopOf="@id/exit"
-            tools:context=".MainActivity">
+            tools:context=".main.MainActivity">
 
         <TableRow
                 android:layout_width="match_parent"
