+++ app/src/main/res/layout/activity_rules.xml
@@ -40,10 +40,22 @@
             android:text="@string/play"
             style="@style/buttonStyle"
             app:layout_constraintTop_toBottomOf="@id/rules_body"
-            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintLeft_toRightOf="@id/auto_play"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"
 
+    />
+
+    <Button
+            android:id="@+id/auto_play"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/auto_play"
+            style="@style/buttonStyle"
+            app:layout_constraintTop_toBottomOf="@id/rules_body"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toLeftOf="@id/play"
+            app:layout_constraintBottom_toBottomOf="parent"
     />
 
 </android.support.constraint.ConstraintLayout>
