+++ app/src/main/res/layout/activity_tie.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@drawable/tttbackground"
-        tools:context=".TieActivity">
+        tools:context=".results.TieActivity">
 
     <ImageView
             android:id="@+id/tie"
