+++ app/src/main/res/values/strings.xml
@@ -18,6 +18,7 @@
     <string name="play">PLAY</string>
     <string name="player2Symbol">O</string>
     <string name="exit">Exit</string>
+    <string name="auto_play">Auto Play</string>
 
 
 </resources>
